Move all non-[start|finish]*State functions to the _helpers.go file
