+++ controllers/nnf_workflow_controller_helpers.go
@@ -1,15 +1,436 @@
 package controllers
 
 import (
+	"context"
 	"fmt"
+	"reflect"
 	"strconv"
 	"strings"
 
 	dwsv1alpha1 "github.com/HewlettPackard/dws/api/v1alpha1"
 	"github.com/HewlettPackard/dws/utils/dwdparse"
+	lusv1alpha1 "github.com/NearNodeFlash/lustre-fs-operator/api/v1alpha1"
 	nnfv1alpha1 "github.com/NearNodeFlash/nnf-sos/api/v1alpha1"
+
+	"github.com/go-logr/logr"
+	corev1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
+
+// Validate the workflow and return any error found
+func (r *NnfWorkflowReconciler) validateWorkflow(ctx context.Context, wf *dwsv1alpha1.Workflow) error {
+
+	var createPersistentCount, deletePersistentCount, directiveCount int
+	for _, directive := range wf.Spec.DWDirectives {
+
+		directiveCount++
+		// The webhook validated directives, ignore errors
+		dwArgs, _ := dwdparse.BuildArgsMap(directive)
+		command := dwArgs["command"]
+
+		switch command {
+		case "jobdw":
+
+		case "copy_in", "copy_out":
+			if err := r.validateStagingDirective(ctx, wf, directive); err != nil {
+				return nnfv1alpha1.NewWorkflowError("Invalid staging Directive: " + directive).WithFatal().WithError(err)
+			}
+
+		case "create_persistent":
+			createPersistentCount++
+
+		case "destroy_persistent":
+			deletePersistentCount++
+
+		case "persistentdw":
+			if err := r.validatePersistentInstanceDirective(ctx, wf, directive); err != nil {
+				return nnfv1alpha1.NewWorkflowError("Could not validate persistent instance: " + directive).WithFatal().WithError(err)
+			}
+		}
+	}
+
+	if directiveCount > 1 {
+		// Ensure create_persistent or destroy_persistent are singletons in the workflow
+		if createPersistentCount+deletePersistentCount > 0 {
+			return nnfv1alpha1.NewWorkflowError("Only a single create_persistent or destroy_persistent directive is allowed per workflow").WithFatal()
+		}
+	}
+
+	return nil
+}
+
+// validateStagingDirective validates the staging copy_in/copy_out directives.
+func (r *NnfWorkflowReconciler) validateStagingDirective(ctx context.Context, wf *dwsv1alpha1.Workflow, directive string) error {
+	// Validate staging directive of the form...
+	//   #DW copy_in source=[SOURCE] destination=[DESTINATION]
+	//   #DW copy_out source=[SOURCE] destination=[DESTINATION]
+
+	// For each copy_in/copy_out directive
+	//   Make sure all $JOB_DW_ references point to job storage instance names
+	//   Make sure all $PERSISTENT_DW references an existing persistentdw instance
+	//   Otherwise, make sure source/destination is prefixed with a valid global lustre file system
+	validateStagingArgument := func(arg string) error {
+		name, _ := splitStagingArgumentIntoNameAndPath(arg)
+		if strings.HasPrefix(arg, "$JOB_DW_") {
+			if findDirectiveIndexByName(wf, name) == -1 {
+				return nnfv1alpha1.NewWorkflowError(fmt.Sprintf("Job storage instance '%s' not found", name)).WithFatal()
+			}
+		} else if strings.HasPrefix(arg, "$PERSISTENT_DW_") {
+			if err := r.validatePersistentInstanceByName(ctx, name, wf.Namespace); err != nil {
+				return nnfv1alpha1.NewWorkflowError(fmt.Sprintf("Persistent storage instance '%s' not found", name)).WithFatal()
+			}
+			if findDirectiveIndexByName(wf, name) == -1 {
+				return nnfv1alpha1.NewWorkflowError(fmt.Sprintf("persistentdw directive mentioning '%s' not found", name)).WithFatal()
+			}
+		} else {
+			if r.findLustreFileSystemForPath(ctx, arg, r.Log) == nil {
+				return nnfv1alpha1.NewWorkflowError(fmt.Sprintf("Global Lustre file system containing '%s' not found", arg)).WithFatal()
+			}
+		}
+
+		return nil
+	}
+
+	args, err := dwdparse.BuildArgsMap(directive)
+	if err != nil {
+		return nnfv1alpha1.NewWorkflowError("Invalid DW directive: " + directive).WithFatal()
+	}
+
+	if err := validateStagingArgument(args["source"]); err != nil {
+		return err
+	}
+
+	if err := validateStagingArgument(args["destination"]); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// validatePersistentInstance validates the persistentdw directive.
+func (r *NnfWorkflowReconciler) validatePersistentInstanceDirective(ctx context.Context, wf *dwsv1alpha1.Workflow, directive string) error {
+	// Validate that the persistent instance is available and not in the process of being deleted
+	args, err := dwdparse.BuildArgsMap(directive)
+	if err != nil {
+		return nnfv1alpha1.NewWorkflowError("Invalid DW directive: " + directive).WithFatal()
+	}
+
+	return r.validatePersistentInstanceByName(ctx, args["name"], wf.Namespace)
+}
+
+// validatePersistentInstance validates the persistentdw directive.
+func (r *NnfWorkflowReconciler) validatePersistentInstanceByName(ctx context.Context, name string, namespace string) error {
+	psi, err := r.getPersistentStorageInstance(ctx, name, namespace)
+	if err != nil {
+		return err
+	}
+
+	if !psi.DeletionTimestamp.IsZero() {
+		return nnfv1alpha1.NewWorkflowError("Persistent storage instance " + name + " is deleting").WithFatal()
+	}
+
+	return nil
+}
+
+// Retrieve the persistent storage instance with the specified name
+func (r *NnfWorkflowReconciler) getPersistentStorageInstance(ctx context.Context, name string, namespace string) (*dwsv1alpha1.PersistentStorageInstance, error) {
+	psi := &dwsv1alpha1.PersistentStorageInstance{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name,
+			Namespace: namespace,
+		},
+	}
+
+	err := r.Get(ctx, client.ObjectKeyFromObject(psi), psi)
+	return psi, err
+}
+
+// generateDirectiveBreakdown creates a DirectiveBreakdown for any #DW directive that needs to specify storage
+// or compute node information to the WLM (jobdw, create_persistent, persistentdw)
+func (r *NnfWorkflowReconciler) generateDirectiveBreakdown(ctx context.Context, dwIndex int, workflow *dwsv1alpha1.Workflow, log logr.Logger) (*dwsv1alpha1.DirectiveBreakdown, error) {
+
+	// DWDirectives that we need to generate directiveBreakdowns for look like this:
+	//  #DW command            arguments...
+	//  --- -----------------  -------------------------------------------
+	// "#DW jobdw              type=raw    capacity=9TB name=thisIsReallyRaw"
+	// "#DW jobdw              type=xfs    capacity=9TB name=thisIsXFS"
+	// "#DW jobdw              type=gfs2   capacity=9GB name=thisIsGfs2"
+	// "#DW jobdw              type=lustre capacity=9TB name=thisIsLustre"
+	// "#DW create_persistent  type=lustre capacity=9TB name=thisIsPersistent"
+	// "#DW persistentdw       name=thisIsPersistent"
+
+	// #DW commands that require a dwDirectiveBreakdown
+	breakDownCommands := []string{
+		"jobdw",
+		"create_persistent",
+		"persistentdw",
+	}
+
+	directive := workflow.Spec.DWDirectives[dwIndex]
+	dwArgs, _ := dwdparse.BuildArgsMap(directive)
+	for _, breakThisDown := range breakDownCommands {
+		// We care about the commands that generate a breakdown
+		if breakThisDown == dwArgs["command"] {
+			dwdName := indexedResourceName(workflow, dwIndex)
+			directiveBreakdown := &dwsv1alpha1.DirectiveBreakdown{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      dwdName,
+					Namespace: workflow.Namespace,
+				},
+			}
+
+			result, err := ctrl.CreateOrUpdate(ctx, r.Client, directiveBreakdown,
+				// Mutate function to fill in a directiveBreakdown
+				func() error {
+					dwsv1alpha1.AddWorkflowLabels(directiveBreakdown, workflow)
+					dwsv1alpha1.AddOwnerLabels(directiveBreakdown, workflow)
+					addDirectiveIndexLabel(directiveBreakdown, dwIndex)
+
+					directiveBreakdown.Spec.Directive = directive
+
+					// Link the directive breakdown to the workflow
+					return ctrl.SetControllerReference(workflow, directiveBreakdown, r.Scheme)
+				})
+
+			if err != nil {
+				log.Error(err, "failed to create or update DirectiveBreakdown", "name", directiveBreakdown.Name)
+				return nil, fmt.Errorf("CreateOrUpdate failed for DirectiveBreakdown %v: %w", client.ObjectKeyFromObject(directiveBreakdown), err)
+			}
+
+			if result == controllerutil.OperationResultCreated {
+				log.Info("Created DirectiveBreakdown", "name", directiveBreakdown.Name)
+			} else if result == controllerutil.OperationResultNone {
+				// no change
+			} else {
+				log.Info("Updated DirectiveBreakdown", "name", directiveBreakdown.Name)
+			}
+
+			// The directive's command has been matched, no need to look at the other breakDownCommands
+			return directiveBreakdown, nil
+		}
+	}
+
+	return nil, nil
+}
+
+func (r *NnfWorkflowReconciler) createNnfStorage(ctx context.Context, workflow *dwsv1alpha1.Workflow, s *dwsv1alpha1.Servers, index int, log logr.Logger) (*nnfv1alpha1.NnfStorage, error) {
+	nnfStorage := &nnfv1alpha1.NnfStorage{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      s.Name,
+			Namespace: s.Namespace,
+		},
+	}
+
+	dwArgs, err := dwdparse.BuildArgsMap(workflow.Spec.DWDirectives[index])
+	if err != nil {
+		return nil, nnfv1alpha1.NewWorkflowError("Invalid DW directive: " + workflow.Spec.DWDirectives[index]).WithFatal()
+	}
+
+	pinnedName, pinnedNamespace := getStorageReferenceNameFromWorkflowActual(workflow, index)
+	nnfStorageProfile, err := findPinnedProfile(ctx, r.Client, pinnedNamespace, pinnedName)
+	if err != nil {
+		log.Error(err, "Unable to find pinned NnfStorageProfile", "name", pinnedName)
+		return nil, fmt.Errorf("Could not find pinned NnfStorageProfile %v: %w", types.NamespacedName{Name: pinnedName, Namespace: pinnedNamespace}, err)
+	}
+
+	var owner metav1.Object = workflow
+	if dwArgs["command"] == "create_persistent" {
+		psi, err := r.findPersistentInstance(ctx, workflow, dwArgs["name"])
+		if err != nil {
+			return nil, fmt.Errorf("Could not find PersistentStorageInstance %v for 'create_persistent' directive: %w", dwArgs["name"], err)
+		}
+
+		owner = psi
+	}
+
+	result, err := ctrl.CreateOrUpdate(ctx, r.Client, nnfStorage,
+		func() error {
+			dwsv1alpha1.AddWorkflowLabels(nnfStorage, workflow)
+			dwsv1alpha1.AddOwnerLabels(nnfStorage, owner)
+			addDirectiveIndexLabel(nnfStorage, index)
+
+			nnfStorage.Spec.FileSystemType = dwArgs["type"]
+
+			// Need to remove all of the AllocationSets in the NnfStorage object before we begin
+			nnfStorage.Spec.AllocationSets = []nnfv1alpha1.NnfStorageAllocationSetSpec{}
+
+			// Iterate the Servers data elements to pull out the allocation sets for the server
+			for i := range s.Spec.AllocationSets {
+				nnfAllocSet := nnfv1alpha1.NnfStorageAllocationSetSpec{}
+
+				nnfAllocSet.Name = s.Spec.AllocationSets[i].Label
+				nnfAllocSet.Capacity = s.Spec.AllocationSets[i].AllocationSize
+				if dwArgs["type"] == "lustre" {
+					nnfAllocSet.NnfStorageLustreSpec.TargetType = strings.ToUpper(s.Spec.AllocationSets[i].Label)
+					nnfAllocSet.NnfStorageLustreSpec.BackFs = "zfs"
+					nnfAllocSet.NnfStorageLustreSpec.FileSystemName = string(s.GetUID())[:8]
+					lustreData := mergeLustreStorageDirectiveAndProfile(dwArgs, nnfStorageProfile)
+					if len(lustreData.ExternalMGS) > 0 {
+						nnfAllocSet.NnfStorageLustreSpec.ExternalMgsNid = lustreData.ExternalMGS
+					}
+				}
+
+				// Create Nodes for this allocation set.
+				for _, storage := range s.Spec.AllocationSets[i].Storage {
+					node := nnfv1alpha1.NnfStorageAllocationNodes{Name: storage.Name, Count: storage.AllocationCount}
+					nnfAllocSet.Nodes = append(nnfAllocSet.Nodes, node)
+				}
+
+				nnfStorage.Spec.AllocationSets = append(nnfStorage.Spec.AllocationSets, nnfAllocSet)
+			}
+
+			// The Servers object owns the NnfStorage object, so it will be garbage collected when the
+			// Server object is deleted.
+			return ctrl.SetControllerReference(owner, nnfStorage, r.Scheme)
+		})
+
+	if err != nil {
+		log.Error(err, "Failed to create or update NnfStorage", "name", nnfStorage.Name)
+		return nnfStorage, fmt.Errorf("CreateOrUpdate failed for NnfStorage %v: %w", client.ObjectKeyFromObject(nnfStorage), err)
+	}
+
+	if result == controllerutil.OperationResultCreated {
+		log.Info("Created NnfStorage", "name", nnfStorage.Name)
+	} else if result == controllerutil.OperationResultNone {
+		// no change
+	} else {
+		log.Info("Updated NnfStorage", "name", nnfStorage.Name)
+	}
+
+	return nnfStorage, nil
+}
+
+func (r *NnfWorkflowReconciler) findLustreFileSystemForPath(ctx context.Context, path string, log logr.Logger) *lusv1alpha1.LustreFileSystem {
+	lustres := &lusv1alpha1.LustreFileSystemList{}
+	if err := r.List(ctx, lustres); err != nil {
+		log.Error(err, "Failed to list lustre file systems")
+		return nil
+	}
+
+	for _, lustre := range lustres.Items {
+		if strings.HasPrefix(path, lustre.Spec.MountRoot) {
+			return &lustre
+		}
+	}
+
+	return nil
+}
+
+func (r *NnfWorkflowReconciler) setupNnfAccessForServers(ctx context.Context, storage *nnfv1alpha1.NnfStorage, workflow *dwsv1alpha1.Workflow, index int, parentDwIndex int, teardownState string, log logr.Logger) (*nnfv1alpha1.NnfAccess, error) {
+
+	params, err := dwdparse.BuildArgsMap(workflow.Spec.DWDirectives[parentDwIndex])
+	if err != nil {
+		return nil, fmt.Errorf("Unable to parse directive %s: %w", workflow.Spec.DWDirectives[parentDwIndex], err)
+	}
+
+	access := &nnfv1alpha1.NnfAccess{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      indexedResourceName(workflow, parentDwIndex) + "-servers",
+			Namespace: workflow.Namespace,
+		},
+	}
+
+	result, err := ctrl.CreateOrUpdate(ctx, r.Client, access,
+		func() error {
+			dwsv1alpha1.AddWorkflowLabels(access, workflow)
+			dwsv1alpha1.AddOwnerLabels(access, workflow)
+			nnfv1alpha1.AddDataMovementTeardownStateLabel(access, teardownState)
+			addDirectiveIndexLabel(access, index)
+
+			access.Spec = nnfv1alpha1.NnfAccessSpec{
+				DesiredState:    "mounted",
+				TeardownState:   teardownState,
+				Target:          "all",
+				MountPath:       buildMountPath(workflow, params["name"], params["command"]),
+				MountPathPrefix: buildMountPath(workflow, params["name"], params["command"]),
+
+				// NNF Storage is Namespaced Name to the servers object
+				StorageReference: corev1.ObjectReference{
+					Kind:      reflect.TypeOf(nnfv1alpha1.NnfStorage{}).Name(),
+					Name:      storage.Name,
+					Namespace: storage.Namespace,
+				},
+			}
+
+			return ctrl.SetControllerReference(workflow, access, r.Scheme)
+		})
+
+	if err != nil {
+		return nil, fmt.Errorf("CreateOrUpdate failed for NnfAccess %v: %w", client.ObjectKeyFromObject(access), err)
+	}
+
+	if result == controllerutil.OperationResultCreated {
+		log.Info("Created NnfAccess", "name", access.Name, "teardown", teardownState)
+	} else if result == controllerutil.OperationResultUpdated {
+		log.Info("Updated NnfAccess", "name", access.Name, "teardown", teardownState)
+	}
+
+	return access, nil
+}
+
+func (r *NnfWorkflowReconciler) getDirectiveFileSystemType(ctx context.Context, workflow *dwsv1alpha1.Workflow, index int) (string, error) {
+	dwArgs, _ := dwdparse.BuildArgsMap(workflow.Spec.DWDirectives[index])
+	switch dwArgs["command"] {
+	case "jobdw":
+		return dwArgs["type"], nil
+	case "persistentdw":
+		name, namespace := getStorageReferenceNameFromWorkflowActual(workflow, index)
+		nnfStorage := &nnfv1alpha1.NnfStorage{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      name,
+				Namespace: namespace,
+			},
+		}
+
+		if err := r.Get(ctx, client.ObjectKeyFromObject(nnfStorage), nnfStorage); err != nil {
+			return "", fmt.Errorf("Could not get persistent NnfStorage %v to determine file system type: %w", client.ObjectKeyFromObject(nnfStorage), err)
+		}
+
+		return nnfStorage.Spec.FileSystemType, nil
+	default:
+		return "", fmt.Errorf("Invalid directive '%s' to get file system type", workflow.Spec.DWDirectives[index])
+	}
+}
+
+// TODO: Can this function be changed to accept a workflow and directive index? That way it can be used
+// in the data movement code setupNnfAccessForServers()
+func buildMountPath(workflow *dwsv1alpha1.Workflow, name string, command string) string {
+
+	if len(name) == 0 {
+		panic(fmt.Sprintf("Mount Path: Empty name"))
+	}
+
+	switch command {
+	case "jobdw":
+		return fmt.Sprintf("/mnt/nnf/%d/job/%s", workflow.Spec.JobID, name)
+	case "persistentdw":
+		return fmt.Sprintf("/mnt/nnf/%d/persistent/%s", workflow.Spec.JobID, name)
+	}
+
+	panic(fmt.Sprintf("Mount Path: Invalid command '%s'", command))
+}
+
+func (r *NnfWorkflowReconciler) findPersistentInstance(ctx context.Context, wf *dwsv1alpha1.Workflow, psiName string) (*dwsv1alpha1.PersistentStorageInstance, error) {
+	log := r.Log.WithValues("Workflow", types.NamespacedName{Name: wf.Name, Namespace: wf.Namespace})
+
+	psi := &dwsv1alpha1.PersistentStorageInstance{}
+	psiNamedNamespace := types.NamespacedName{Name: psiName, Namespace: wf.Namespace}
+	err := r.Get(ctx, psiNamedNamespace, psi)
+	if err != nil {
+		if !apierrors.IsNotFound(err) {
+			log.Error(err, "Unable to get PersistentStorageInstance", "name", psiName, "error", err)
+		}
+
+		return nil, err
+	}
+
+	return psi, err
+}
 
 func handleWorkflowError(err error, driverStatus *dwsv1alpha1.WorkflowDriverStatus) {
 	e, ok := err.(*nnfv1alpha1.WorkflowError)
